Implement compatibility with MoreCompany cosmetics
To allow the local player to see their own cosmetics in the body cam, a patch is applied to MoreCompany to remove a check that prevents cosmetics from being spawned onto the local player.

Cosmetics will be forcibly hidden or displayed based on the perspective of the body cam to emulate that the main camera is viewing whatever players are present.