Keep a reference to the material and a night vision light prefab always
To ensure that we always have access to these to recreate the camera, grab them when the bodycam is created and keep them around.

This fixes a compatibility issue with Immersive Visor, which would destroy the visor and its children, including the plane that is used to apply an extra fog shader to the game.