Keep the main map screen rendering when the terminal is in use
The internal ship camera rendering the monitors at the front of the ship keeps the main map screen from being culled, therefore keeping the map rendering while the player is in the terminal. However, with the DisableInternalShipCamera enabled, that is no longer the case, so we have to specifically ensure that the map camera remains enabled.