Keep the external camera enabled when the screen by the door is visible
Since the internal camera no longer forces the external camera to stay enabled, we have to add a check for the visibility of the screen by the door to allow that to render while the player is looking at it.

This check uses a manual frustum check instead of `Renderer.isVisible` to avoid including shadow casting in the visibility check.