Tentative fix for a screen freeze when viewing a masked enemy on cam
It seems as though for some users a renderer attached to a masked enemy will be deleted while a reference is still held by both the MaskedPlayerEnemy instance and our body cam. To avoid this triggering a failed frame render, add a null check before forcing any renderers to stop rendering.