React when ModelReplacementAPI changes the player model replacement
Previously, we were only collecting the models used to replace the normal player model based on unrelated events, which meant that if the model was removed or changed after the models were referenced by the body cam, error spam would occur.

To avoid this, hook into `ViewStateManager.ReportBodyReplacementAddition()` to be notified when the model is changed. Since the method to change the model replacement is to `Object.Destroy()` a `BodyReplacementBase` component and then create a new one, and the old body replacement stays alive until the end of the frame due to deferred destruction, the private field `ViewStateManager.bodyReplacement` is used instead of `GetComponent()` to get the current model replacer.