Exclude body cams' children from `currentlyViewedMeshes`
Previously, the renderer for the green flash transition animation would be included in this array when viewing dead bodies or masked enemies. This would cause the green flash to be treated as a third-person model, hidden in first person and displayed in third person.

Fixes the bug where switching from a dead body to a player would briefly or permanently display the green flash on the gameplay camera.