Make null checks in `currentlyViewedMeshes` consistent
Previously, `BeginCameraRendering()` would check and skip nulls, whereas `ResetCameraRendering()` would skip all renderers if the first one was null. This could cause meshes to remain hidden when they should be reverted to be visible.

An error message will be printed when nulls are found in `BeginCameraRendering()`, but will be ignored in `ResetCameraRendering()`, since nothing should be destroying the renderers between the two.