Enable backface culling on ragdolls in body cams
In v55, the ragdoll model began using a double-sided material, which causes the body cam to become obstructed. We have to correct this when the camera is attached to a ragdoll.

Note that we use the shared material to avoid duplicating the materials, which will mean that all ragdolls sharing a mterial will be backface-culled in the body cam. This may cause some visual inconsistency in some cases.