Keep HDRP camera data persistent for body cams
When a camera is disabled for a frame, HDRP will delete its data associated with it, meaning that it will have to set it up again on any subsequent frames where it is enabled. This causes a fairly significant overhead when running a framerate limit on body cams.

This has the side effect of keeping information for the custom pass around for all body cams, so the flickering in the custom pass is no longer present.

Fixes #29