Hook into HDRenderPipeline for camera callbacks
The callbacks that exist apparently aren't called at the appropriate time to be able to adjust certain rendering parameters.

For example, light settings are not calculated during the camera culling loop, but rather the camera rendering loop that follows it. This also probably applies to material options.

Also, the volume weights are calculated during the culling loop, but before the beginCameraRendering callback is invoked.

A helper class to find injection points and inject code has also been added.