Add a patch to allow clients' maps to target converted masked players
All clients wait to be able to delete the converted player's dead body, but the kill animation does not actually spawn a dead body, so it gets stuck and never assigns the player's redirect target.

To avoid this, the patch just makes the kill animation spawn a body that is then immediately deactivated.