Convert the MoreCompany local cosmetics patch to a transpiler
This allows us to hopefully react properly to any changes to the underlying logic, or if it changes to significantly, to break the patch instead of breaking the cosmetics.

Updating the cosmetics is still done in a postfix so that that patch can function even if patching to spawn local cosmetics fails.