Make the weather effects at the target of a body cam always visible
The weather effects will be cloned for each body cam, and follow their targets. When the camera is rendered, the vanilla ones are hidden and the clones are shown.

This allows effects like the rolling dust clouds and rain to appear on the body cam even when the viewer is in an area that lacks the weather effect that the viewed player can see.